git-svn-id: svn+ssh://svn.code.sf.net/p/swistrack/code@291 18c3c8b4-d610-0410-8eed-94a590f67c64

+++ code-branch-thomas/core/RandomNormal.cpp
@@ -50,7 +50,7 @@ double THISCLASS::nfix(long hz, unsigned long iz) {
 
 		hz=mRandomMersenneTwister->randInt();
 		iz=hz & 127;
-		if (abs(hz)<kn[iz]) {
+		if ((unsigned long)abs(hz)<kn[iz]) {
 			return (hz*wn[iz]);
 		}
 	}
